Moving PageLoading into “generic” sub-directory.
