+++ src/courseware/course/sequence/SequenceContent.jsx
@@ -1,7 +1,7 @@
 import React, { Suspense, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
-import PageLoading from '../../../PageLoading';
+import PageLoading from '../../../generic/PageLoading';
 import { useModel } from '../../../generic/model-store';
 
 import messages from './messages';
